refactor: rename deprecation-map to map-deprecated
